fix: return a float from the BHT-002-GCLZB calibration fix
math.ceil and math.floor answer with an int, so this was the one
implementation of fix_local_calibration whose result did not match the
float its signature and docstring declare.

+++ custom_components/better_thermostat/model_fixes/BHT-002-GCLZB.py
@@ -34,9 +34,9 @@ def fix_local_calibration(self: ModelFixHost, entity_id: str, offset: float) -> 
         The normalized integer-valued calibration offset.
     """
     if self.cur_temp < self.bt_target_temp:
-        offset = math.ceil(offset)
+        offset = float(math.ceil(offset))
     else:
-        offset = math.floor(offset)
+        offset = float(math.floor(offset))
 
     return offset
